fix: 调整一些代码

+++ src/main/java/cn/bitterfree/sse/util/SseMessageSender.java
@@ -2,17 +2,17 @@ package cn.bitterfree.sse.util;
 
 import cn.bitterfree.common.enums.GlobalServiceStatusCode;
 import cn.bitterfree.common.util.convert.JsonUtil;
+import cn.bitterfree.common.util.juc.threadpool.ThreadPoolUtil;
 import cn.bitterfree.sse.session.SseSessionMapper;
 import io.jsonwebtoken.lang.Collections;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.MediaType;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 
-import java.io.IOException;
-import java.util.List;
-import java.util.Objects;
-import java.util.function.Consumer;
+import java.util.Set;
+import java.util.concurrent.ThreadPoolExecutor;
 import java.util.function.Function;
+import java.util.function.Supplier;
 
 /**
  * Created With Intellij IDEA
@@ -24,45 +24,41 @@ import java.util.function.Function;
 @Slf4j
 public class SseMessageSender {
 
-    private static Consumer<IOException> handleException(String sessionKey) {
-        return e -> {
-            log.error("{} 发送消息异常 {}", sessionKey, e.getMessage());
-            SseSessionMapper.remove(sessionKey);
-        };
-    }
+    private final static ThreadPoolExecutor EXECUTOR = ThreadPoolUtil.getIoTargetThreadPool("SSE-Thread");;
 
-    private static <T> void sendMessage(SseEmitter sseEmitter, T message, Consumer<IOException> handleException) {
-        if(Objects.isNull(sseEmitter)) {
-            log.warn(GlobalServiceStatusCode.SSE_CONNECTION_NOT_EXIST.toString());
-            return;
-        }
-        try {
-            sseEmitter.send(JsonUtil.toJson(message), MediaType.APPLICATION_JSON);
-        } catch (IOException e) {
-            handleException.accept(e);
-        }
+    public static <T> void sendMessage(String sessionKey, Supplier<T> messageSupplier) {
+        EXECUTOR.submit(() -> {
+            T message = messageSupplier.get();
+            log.info("服务器 -> [{}] text: {}", sessionKey, message);
+            if(!SseSessionMapper.containsKey(sessionKey)) {
+                log.warn(GlobalServiceStatusCode.SSE_CONNECTION_NOT_EXIST.getMessage());
+                return;
+            }
+            SseEmitter sseEmitter = SseSessionMapper.get(sessionKey);
+            try {
+                sseEmitter.send(JsonUtil.toJson(message), MediaType.APPLICATION_JSON);
+            } catch (Exception e) {
+                log.error("{} 发送消息异常 {}", sessionKey, e.getMessage());
+                SseSessionMapper.remove(sessionKey);
+            }
+        });
     }
 
     public static <T> void sendMessage(String sessionKey, T message) {
-        log.info("服务器 -> [{}] text: {}", sessionKey, message);
-        SseSessionMapper.consumeKey(sessionKey, sseEmitter -> {
-            sendMessage(sseEmitter, message, handleException(sessionKey));
-        });
+        sendMessage(sessionKey, () -> message);
     }
 
-    public static <T> void sendAllMessage(String prefix, Function<String, T> function) {
-        SseSessionUtil.getSessionKeys(prefix).stream().parallel().forEach(sessionKey -> {
-            sendMessage(sessionKey, function.apply(sessionKey));
-        });
-    }
-
-    public static <T> void sendAllMessage(List<String> sessionKeys, Function<String, T> function) {
+    public static <T> void sendAllMessage(Set<String> sessionKeys, Function<String, T> function) {
         if (Collections.isEmpty(sessionKeys)) {
             return;
         }
         sessionKeys.stream().parallel().distinct().forEach(sessionKey -> {
             sendMessage(sessionKey, function.apply(sessionKey));
         });
+    }
+
+    public static <T> void sendAllMessage(String prefix, Function<String, T> function) {
+        sendAllMessage(SseSessionMapper.getKeys(prefix), function);
     }
 
 }
